requirements.txt
openai
pinecone-client
datasets